+++ src/main/java/com/krystianrymonlipinski/algorithm/playingalgorithm/PositionState.java
@@ -5,12 +5,12 @@ import draughts.library.managers.GameEngine;
 
 public class PositionState {
 
-    private double rewardFunctionOutcome;
+    private float rewardFunctionOutcome;
     private GameEngine.GameState gameState;
     private DrawArbiter.DrawConditions drawConditions;
     private int drawCounter;
-    private double alpha = RewardCalculator.FITNESS_FUNCTION_VALUE_BLACK_WON;
-    private double beta = RewardCalculator.FITNESS_FUNCTION_VALUE_WHITE_WON;
+    private float alpha = RewardCalculator.FITNESS_FUNCTION_VALUE_BLACK_WON;
+    private float beta = RewardCalculator.FITNESS_FUNCTION_VALUE_WHITE_WON;
 
     public PositionState(GameEngine gameEngine) {
         this.gameState = gameEngine.getGameState();
@@ -18,11 +18,11 @@ public class PositionState {
         this.drawCounter = gameEngine.getDrawArbiter().getDrawCounter();
     }
 
-    public double getRewardFunctionOutcome() {
+    public float getRewardFunctionOutcome() {
         return rewardFunctionOutcome;
     }
 
-    public void setRewardFunctionOutcome(double rewardFunctionOutcome) {
+    public void setRewardFunctionOutcome(float rewardFunctionOutcome) {
         this.rewardFunctionOutcome = rewardFunctionOutcome;
     }
 
@@ -50,19 +50,19 @@ public class PositionState {
         this.drawCounter = drawCounter;
     }
 
-    public double getAlpha() {
+    public float getAlpha() {
         return alpha;
     }
 
-    public void setAlpha(double alpha) {
+    public void setAlpha(float alpha) {
         this.alpha = alpha;
     }
 
-    public double getBeta() {
+    public float getBeta() {
         return beta;
     }
 
-    public void setBeta(double beta) {
+    public void setBeta(float beta) {
         this.beta = beta;
     }
 }
